remove device ID handling
